Add crate features to control which model formats are available
Both `.rten` and `.onnx` formats are enabled by default. Disabling unused
formats can improve build times. Enabling both formats generates ~912K LLVM
lines for the rten crate. Enabling only one format generates ~840K lines (-8%).
